e2e: numacell: move logic in packages
Trivial code movement to make it easier to integrate
the numacell plugin into external projects (e.g. cnf-tests)
which also want to integrate the serial suite, which
depends on numacell.

No intended changes on behaviour.
